+++ .pre-commit-config.yaml
@@ -16,3 +16,10 @@ repos:
     hooks:
       - id: interrogate
         args: [ -vv, -i, --fail-under=80 ]
+  - repo: https://github.com/python-poetry/poetry
+    rev: '1.2.2'
+    hooks:
+      - id: poetry-check
+      - id: poetry-lock
+      - id: poetry-export
+        args: [ "-f", "requirements.txt", "-o", "requirements.txt" ]
