+++ README.md
@@ -1,33 +1,45 @@
-# Job Application Tracker
+# Track Job Applications (CLI)
 
-Job application tracker is a CLI tool to help
+Track Job Applications is a CLI app to help
 professionals currently applying for jobs, by keeping
 a detailed record of all the job openings they have
-applied to.
+applied to and provide relevant metrics on them.
 
 ![MIT License](https://img.shields.io/github/license/Aditya-Gupta1/job-application-cli?color=green&style=flat-square)
 ![Open Issues](https://img.shields.io/github/issues/Aditya-Gupta1/job-application-cli?color=dark-green&style=flat-square)
 ![Good First Issues](https://img.shields.io/github/issues/Aditya-Gupta1/job-application-cli/good%20first%20issue?color=blue&style=flat-square)
 ![Hacktoberfest Issues](https://img.shields.io/github/issues/Aditya-Gupta1/job-application-cli/hacktoberfest?color=cyan&style=flat-square)
-![Interrogate](interrogate_badge.svg)
 
 ## Motivation
 
+The question is: **"Why do we even need to track our job applications?"**
+
 We apply to hundreds of job openings when looking out for jobs, and often forget about
 them at a later point in time. Since we forget about them, we can't actually analyze
-what is going right or wrong in the entire process.
+what is going right or wrong in the entire process. 
 
-So the question is: **"Why do we even need to track them?"**
+And actually there's a lot we can do with this data! It can be used to answer questions like:
 
-Well...for any reason!
-
-- You may want to know from which companies you're getting interview calls from.
-- Or, you may want to know where you're not able to get shortlisted.
-- Or, in how many companies you've completed more than 2 rounds of interviews.
-- Or, in how many HR rounds you got rejected.
-- or, how many offers you have. *Though you won't forget this ;)*
+- In how many companies have I been shortlisted yet?
+- What companies rejected my profile?
+- In how many companies have I given 2 rounds of tech interviews?
+- In how many HR rounds I got rejected?
+- How many offers do I have? *Though you won't forget this :wink: )*
 
 And the list goes on.
+
+## Installation
+
+You need to have [Python](https://www.python.org/downloads/) installed on your system as a prerequisite.
+If it is already installed, run the following command:
+
+```cmd
+pip install track-job-applications
+```
+
+## Quick Start
+
+
 
 ## Tech Stack
 
@@ -39,6 +51,6 @@ And the list goes on.
 
 Contributions are always welcome!
 
-See `CONTRIBUTING.md` for ways to get started.
+See [CONTRIBUTING.md](https://github.com/Aditya-Gupta1/track-job-applications/blob/main/CONTRIBUTING.md) for ways to get started.
 
-Please adhere to this project's `CODE OF CONDUCT`.
+Please adhere to this project's [CODE OF CONDUCT](https://github.com/Aditya-Gupta1/track-job-applications/blob/main/CODE_OF_CONDUCT.md).
