+++ pyproject.toml
@@ -33,3 +33,44 @@ whitelist-regex = []
 color = true
 generate-badge = "."
 badge-format = "svg"
+
+[tool.poetry]
+name = "track-job-applications"
+version = "0.0.3"
+description = "A CLI application to track job applications and provide relevant metrics on them."
+authors = ["Aditya Gupta <guptaaditya008@gmail.com>"]
+maintainers = ["Aditya Gupta <guptaaditya008@gmail.com>"]
+repository = "https://github.com/Aditya-Gupta1/track-job-applications"
+license = "MIT"
+readme = "README.md"
+packages = [
+    {include = "track"}
+]
+
+[tool.poetry.dependencies]
+python = "^3.8"
+typer = "^0.6.1"
+click = "^8.0.0"
+SQLAlchemy = "^1.4.41"
+rich = "^12.6.0"
+
+[tool.poetry.group.dev.dependencies]
+typer = "^0.6.1"
+click = "^8.0.0"
+SQLAlchemy = "^1.4.41"
+rich = "^12.6.0"
+pre-commit = "^2.20.0"
+black = "^22.10.0"
+isort = "^5.10.1"
+interrogate = "^1.5.0"
+flake8 = "^5.0.4"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+track-job = 'track.main:entry'
+
+[tool.poetry.urls]
+"Bug Tracker" = "https://github.com/Aditya-Gupta1/track-job-applications/issues"
