+++ track/__main__.py
@@ -1,0 +1,3 @@
+from main import entry
+
+entry()
